[Code cleanup #29] Cleaning up the code with new architecture.
