+++ config/receiver.ini
@@ -23,7 +23,7 @@ data_size = 8
 is_complex = true
 
 [SATELLITES]
-include_prn = 2,3,4,6,9,29
+include_prn = 2,3,4,6,9,11,29,31
 
 [MEASUREMENTS]
 # In Hz
