+++ core/old/acquisition/acquisition_abstract.py
