+++ core/old/decoding/message_abstract.py
