+++ core/old/visualisation.py
@@ -15,7 +15,7 @@ import pymap3d as pm
 from core.utils.constants import SPEED_OF_LIGHT
 from core.receiver.receiver_abstract import ReceiverAbstract
 from core.signal.rfsignal import RFSignal
-from core.satellite.satellite import Satellite
+from core.space.satellite import Satellite
 from core.utils.enumerations import GNSSSignalType
 from core.channel.channel_abstract import ChannelState
